Correct my float errors

+++ db/migrate/20241203194353_create_product_prices.rb
@@ -1,7 +1,7 @@
 class CreateProductPrices < ActiveRecord::Migration[7.1]
   def change
     create_table :product_prices do |t|
-      t.integer :price
+      t.float :price
       t.date :scraping_date
       t.references :product, null: false, foreign_key: true
